library_backend/requirements.txt
# ============================================
# ЗАВИСИМОСТИ BACKEND API LibriMomsClub
# ============================================

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.23
aiosqlite==0.19.0

# JWT и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Валидация
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP клиент (для проверки Telegram данных)
httpx==0.25.2

# Утилиты
python-slugify==8.0.1
